requirements.txt
python-dotenv~=1.0.0
anthropic~=0.25.7